Change learnign rule to prevent large weights
new: weight_change = LEARNING_RATE * activity * reward * fabs(THRESHOLD + 0.5 - fabs(weight));

+++ C/PythonModule/Synapse.h
@@ -5,8 +5,8 @@
 struct Synapse{
   float weight;
   float activity;
-  float input;
-  float *p_presynaptic_output;
+  int input;
+  int *p_presynaptic_output;
 };
